demo-projects/birthie/requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.3.3
python-dotenv==0.20.0
Werkzeug==2.2.2
